fix: track loop/function states inorder to validate break/continue statement
When a while/for loop is encountered, the loop_flag value is added (saved) to the loop_flag_stack and is set to 1.

When a function definition is encountered, the loop_flag value is added to the loop_flag_stack and is set to 0.

When the program exits a for/while loop or a function definition, the previous loop_flag value is retrieved from the loop_flag_stack via pop.